最新の資材になります

+++ 2306teamdevelopment/src/main/java/com/example/demo/repository/ExpenseRepository.java
@@ -7,7 +7,7 @@ import com.example.demo.dto.ExpenseRequest;
 import com.example.demo.entity.ExpenseEntity;
 
 @Repository
-public interface ExpenseRepository extends JpaRepository<ExpenseEntity, Long> {
+public interface ExpenseRepository extends JpaRepository<ExpenseEntity, Integer> {
 
 	void save(ExpenseRequest expenserequest);
